Updates to InputImage
